requirements.txt
pandas==2.0.3
robotframework==7.3.1
robotframework-seleniumlibrary==6.7.1
robotframework-requests==0.9.7
robotframework-pythonlibcore==4.4.1
